Refactor function to send inference and migrate drift test to kserve
