+++ tests/model_explainability/trustyai_service/drift/model_data/data_batches/0.json
@@ -1,41 +1,39 @@
-{
-  "inputs": [
-    {
-      "name": "credit_inputs",
-      "shape": [5, 4],
-      "datatype": "FP64",
-      "data": [
-        [
-          50.76899726547295,
-          491.57388556551217,
-          11.464223840747968,
-          28.29901088554935
-        ],
-        [
-          47.80397343448882,
-          515.4934209257955,
-          13.060710933476372,
-          23.710220802886678
-        ],
-        [
-          56.731646594370886,
-          441.00792486531225,
-          10.616678496549381,
-          19.040822238191925
-        ],
-        [
-          38.35835814888362,
-          540.3557103621482,
-          11.206492946288046,
-          15.561855179575819
-        ],
-        [
-          47.422935471130046,
-          529.5980108579066,
-          10.625654599760802,
-          16.264208531245814
-        ]
+[
+  {
+    "name": "credit_inputs",
+    "shape": [5, 4],
+    "datatype": "FP64",
+    "data": [
+      [
+        50.76899726547295,
+        491.57388556551217,
+        11.464223840747968,
+        28.29901088554935
+      ],
+      [
+        47.80397343448882,
+        515.4934209257955,
+        13.060710933476372,
+        23.710220802886678
+      ],
+      [
+        56.731646594370886,
+        441.00792486531225,
+        10.616678496549381,
+        19.040822238191925
+      ],
+      [
+        38.35835814888362,
+        540.3557103621482,
+        11.206492946288046,
+        15.561855179575819
+      ],
+      [
+        47.422935471130046,
+        529.5980108579066,
+        10.625654599760802,
+        16.264208531245814
       ]
-    }
-  ]
-}
+    ]
+  }
+]
