+++ tests/model_explainability/trustyai_service/drift/model_data/data_batches/5.json
@@ -1,44 +1,42 @@
-{
-  "inputs": [
-    {
-      "name": "credit_inputs",
-      "shape": [
-        5,
-        4
+[
+  {
+    "name": "credit_inputs",
+    "shape": [
+      5,
+      4
+    ],
+    "datatype": "FP64",
+    "data": [
+      [
+        47.10141440274886,
+        511.93953118897673,
+        10.30958871700256,
+        28.462273005138734
       ],
-      "datatype": "FP64",
-      "data": [
-        [
-          47.10141440274886,
-          511.93953118897673,
-          10.30958871700256,
-          28.462273005138734
-        ],
-        [
-          55.018917461242545,
-          529.5655000531016,
-          10.657507738326363,
-          20.254038773880144
-        ],
-        [
-          43.36735715813185,
-          459.45376201509356,
-          11.974670802162198,
-          16.815021767153233
-        ],
-        [
-          48.20533085890888,
-          484.26866990656447,
-          9.765379302729444,
-          20.95457742333733
-        ],
-        [
-          44.52555863555142,
-          448.30189185655723,
-          12.468831395634185,
-          30.501275682394212
-        ]
+      [
+        55.018917461242545,
+        529.5655000531016,
+        10.657507738326363,
+        20.254038773880144
+      ],
+      [
+        43.36735715813185,
+        459.45376201509356,
+        11.974670802162198,
+        16.815021767153233
+      ],
+      [
+        48.20533085890888,
+        484.26866990656447,
+        9.765379302729444,
+        20.95457742333733
+      ],
+      [
+        44.52555863555142,
+        448.30189185655723,
+        12.468831395634185,
+        30.501275682394212
       ]
-    }
-  ]
-}
+    ]
+  }
+]
