+++ tests/model_explainability/trustyai_service/fairness/test_fairness.py
@@ -1,3 +1,5 @@
+import time
+
 import pytest
 
 
@@ -5,11 +7,11 @@ import pytest
     "model_namespace",
     [
         pytest.param(
-            {"name": "test-fairness-loan-model", "modelmesh-enabled": True},
+            {"name": "test-fairness"},
         )
     ],
     indirect=True,
 )
 class TestFairnessMetrics:
-    def test_dummy(self):
-        pass
+    def test_dummy(self, admin_client, model_namespace, trustyai_service_with_pvc_storage, onnx_loan_model):
+        time.sleep(1200)  # noqa: FCN001
