+++ tests/model_serving/model_runtime/vllm/conftest.py
@@ -12,9 +12,10 @@ from utilities.constants import KServeDeploymentType
 from pytest import FixtureRequest
 from syrupy.extensions.json import JSONSnapshotExtension
 from tests.model_serving.model_runtime.vllm.utils import get_runtime_manifest
-from tests.model_serving.model_server.utils import create_isvc
 from tests.model_serving.model_runtime.vllm.constant import TEMPLATE_MAP, ACCELERATOR_IDENTIFIER, PREDICT_RESOURCES
 from simple_logger.logger import get_logger
+
+from utilities.infra import create_isvc
 from utilities.infra import get_pods_by_isvc_label
 
 LOGGER = get_logger(name=__name__)
