+++ tests/model_serving/model_server/components/conftest.py
@@ -12,7 +12,7 @@ from ocp_resources.serving_runtime import ServingRuntime
 
 from utilities.data_science_cluster_utils import update_components_in_dsc
 from utilities.constants import DscComponents
-from tests.model_serving.model_server.utils import create_isvc
+from utilities.infra import create_isvc
 
 
 @pytest.fixture(scope="class")
