+++ tests/model_serving/model_server/conftest.py
@@ -16,7 +16,6 @@ from ocp_resources.serving_runtime import ServingRuntime
 from ocp_resources.storage_class import StorageClass
 from pytest_testconfig import config as py_config
 
-from tests.model_serving.model_server.utils import create_isvc
 from utilities.constants import DscComponents, StorageClassName
 from utilities.constants import (
     KServeDeploymentType,
@@ -27,7 +26,7 @@ from utilities.constants import (
     Protocols,
     RuntimeTemplates,
 )
-from utilities.infra import s3_endpoint_secret
+from utilities.infra import s3_endpoint_secret, create_isvc
 from utilities.data_science_cluster_utils import update_components_in_dsc
 from utilities.serving_runtime import ServingRuntimeFromTemplate
