+++ tests/model_serving/model_server/model_mesh/conftest.py
@@ -9,7 +9,6 @@ from ocp_resources.secret import Secret
 from ocp_resources.service_account import ServiceAccount
 from ocp_resources.serving_runtime import ServingRuntime
 
-from tests.model_serving.model_server.utils import create_isvc
 from utilities.constants import (
     KServeDeploymentType,
     ModelFormat,
@@ -17,6 +16,7 @@ from utilities.constants import (
     Protocols,
     RuntimeTemplates,
 )
+from utilities.infra import create_isvc
 from utilities.serving_runtime import ServingRuntimeFromTemplate
